+++ packages-node/providence-analytics/src/program/core/QueryService.js
@@ -107,20 +107,27 @@ export class QueryService {
    * @param {AnalyzerAst} requiredAst
    */
   static async addAstToProjectsData(projectsData, requiredAst) {
-    return projectsData.map(projectData => {
+    const resultWithAsts = [];
+
+    for (const projectData of projectsData) {
       const cachedData = astProjectsDataCache.get(projectData.project.path);
       if (cachedData) {
         return cachedData;
       }
 
-      const resultEntries = projectData.entries.map(entry => {
-        const ast = AstService.getAst(entry.context.code, requiredAst, { filePath: entry.file });
-        return { ...entry, ast };
-      });
+      const resultEntries = [];
+      for (const entry of projectData.entries) {
+        const ast = await AstService.getAst(entry.context.code, requiredAst, {
+          filePath: entry.file,
+        });
+        resultEntries.push({ ...entry, ast });
+      }
       const astData = { ...projectData, entries: resultEntries };
       this._addToProjectsDataCache(`${projectData.project.path}#${requiredAst}`, astData);
-      return astData;
-    });
+      resultWithAsts.push(astData);
+    }
+
+    return resultWithAsts;
   }
 
   /**